requirements.txt
cryptography
pycryptodome
streamlit
firebase-admin
python-dotenv

gAAAAABl-ZlMHOeovTpiOgR7KsHzfOgxlGjGjAKQU9Gq74exiT-aUnXTHW_BrLyvWzly8-MAiz8wn8WFIXPUsPZ2kvTE_mUScQ==